build : --Added the separated components and moved into folder
