+++ packages/gatsby-theme-aio/package.json
