+++ packages/gatsby-theme-aio/src/components/GetCredential/Card/CardAPIKey.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import { css } from "@emotion/react";
+import ShowCard from './ShowCard';
+
+const CardAPIKey = ({ cardClientDetails, cardAPIKey, apiKey }) => {
+  return (
+    <ShowCard heading={cardAPIKey?.heading} value={apiKey} />
+  )
+}
+
+export { CardAPIKey };
