+++ packages/gatsby-theme-aio/src/components/GetCredential/Card/CardAllowedOrigins.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import ShowCard from './ShowCard';
+
+const CardAllowedOrigins = ({ cardClientDetails, cardAllowedOrigins, allowedOrigins }) => {
+  return (
+    <ShowCard heading={cardAllowedOrigins?.heading} value={allowedOrigins} />
+  )
+}
+
+
+export { CardAllowedOrigins };
