+++ packages/gatsby-theme-aio/src/components/GetCredential/Card/CardClientId.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import ShowCard from './ShowCard'
+
+const CardClientId = ({ cardClientDetails, cardClientId, apiKey }) => {
+  return (
+    <ShowCard heading={cardClientId?.heading} value={apiKey} />
+  )
+}
+
+export { CardClientId }
