+++ packages/gatsby-theme-aio/src/components/GetCredential/Card/CardClientSecret.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import ShowCard from './ShowCard';
+
+const CardClientSecret = ({ cardClientDetails, cardClientSecret }) => {
+  return (
+    <ShowCard heading={cardClientSecret?.heading} isClientSecret buttonLabel={cardClientSecret?.buttonLabel} />
+  )
+}
+
+
+export { CardClientSecret };
