+++ packages/gatsby-theme-aio/src/components/GetCredential/Card/CardOrganizationName.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import ShowCard from './ShowCard';
+
+const CardOrganizationName = ({ cardClientDetails, cardOrganizationName, organization }) => {
+  return (
+    <ShowCard heading={cardOrganizationName?.heading} value={organization} isOraganization />
+  )
+}
+
+export { CardOrganizationName };
