+++ packages/gatsby-theme-aio/src/components/GetCredential/Card/CardScopes.js
@@ -1,0 +1,9 @@
+import React from 'react'
+import ShowCard from './ShowCard';
+
+const CardScopes = ({ cardClientDetails, cardScopes, allowedOrigins }) => {
+  return (
+    <ShowCard heading={cardScopes?.heading} value={allowedOrigins} />
+  )
+}
+export { CardScopes };
