+++ packages/gatsby-theme-aio/src/components/GetCredential/Card/MyCredentialSide.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const MyCredentialSide = ({ side }) => (side);
+
+export { MyCredentialSide }
