+++ packages/gatsby-theme-aio/src/components/GetCredential/Form/SideCredential.js
@@ -1,0 +1,3 @@
+const SideCredential = ({ side }) => (side);
+
+export { SideCredential };
