+++ packages/gatsby-theme-aio/src/components/GetCredential/FormFields.js
@@ -149,11 +149,12 @@ export const getOrganization = async (setOrganizationValue) => {
         headers: {
           "Content-Type": "application/json",
           "Authorization": "Bearer " + token,
-          "x-api-key": "stage_adobe_io"
+          "x-api-key": window?.adobeIMS?.adobeIdData?.client_id
         }
       });
 
       const organization = await response.json();
+      console.log('organization-------', organization)
 
       if (setOrganizationValue) {
         setOrganizationValue(organization[0]);
