+++ packages/gatsby-theme-aio/src/components/GetCredential/MyCredential.js
@@ -4,10 +4,20 @@ import classNames from "classnames";
 import JSZip from 'jszip';
 import JSZipUtils from 'jszip-utils';
 import { saveAs } from 'file-saver';
-import { SideContent } from './CredentialForm';
-import { CopyIcon, KeyIcon, LinkOut, MAX_TABLET_SCREEN_WIDTH, MIN_MOBILE_WIDTH } from './FormFields';
+import { KeyIcon, LinkOut, MAX_TABLET_SCREEN_WIDTH, MIN_MOBILE_WIDTH } from './FormFields';
 import { Toast } from '../Toast';
 import { CardProducts } from './Products';
+import { MyCredentialSide } from './Card/MyCredentialSide';
+import { AccessToken } from './Card/AccessToken';
+import { DevConsoleLink } from './Card/DevConsoleLink';
+import { CardClientDetails } from './Card/CardClientDetails';
+import { CardClientId } from './Card/CardClientId';
+import { CardClientSecret } from './Card/CardClientSecret';
+import { CardOrganizationName } from './Card/CardOrganizationName';
+import { CardScopes } from './Card/CardScopes';
+import { CardAPIKey } from './Card/CardAPIKey';
+import { CardAllowedOrigins } from './Card/CardAllowedOrigins';
+import { SideContent } from './Form/SideContent';
 
 const MyCredential = ({
   credentialProps,
@@ -425,204 +435,4 @@ const MyCredential = ({
 
 }
 
-const MyCredentialSide = ({ side }) => (side);
-
-const AccessToken = ({ accessToken }) => {
-  return (
-    <>
-      {accessToken && <div css={css`
-        display : flex;
-        flex-direction : column;
-        gap:16px;
-      `}>
-        {accessToken?.heading && <h4 className="spectrum-Heading spectrum-Heading--sizeS">{accessToken?.heading}</h4>}
-        {accessToken?.buttonLabel && <button css={css`width: 180px;`} className="spectrum-Button spectrum-Button--fill spectrum-Button--accent spectrum-Button--sizeM">
-          <span className="spectrum-Button-label">{accessToken?.buttonLabel}</span>
-        </button>}
-      </div>}
-    </>
-  )
-};
-
-const DevConsoleLink = ({ cardDevConsoleLink, formData }) => {
-  return (
-    <>
-      <div css={css`
-          display : flex;
-          flex-direction : column;
-          gap:16px;
-        `}>
-        {cardDevConsoleLink?.heading && <h4 className="spectrum-Heading spectrum-Heading--sizeS">{cardDevConsoleLink?.heading}</h4>}
-
-        <div css={css`display:flex;`}>
-          <div><p className="spectrum-Body spectrum-Body--sizeS"
-            css={css`
-              font-family: Source Code Pro,Monaco,monospace;
-              white-space: normal;
-              overflow-wrap: anywhere;
-              max-width: 300px;
-              color: #0265DC;
-            `}
-          >{formData['CredentialName']}</p></div>
-          <div css={
-            css`
-              margin-left:10px;
-              @media screen and (min-width:${MIN_MOBILE_WIDTH}) and (max-width:${MAX_TABLET_SCREEN_WIDTH}){
-                display:none;
-              }
-            }`
-          }><LinkOut /></div>
-        </div>
-      </div>
-    </>
-  )
-}
-
-const CardClientDetails = ({ cardClientDetails, cardClientId, cardClientSecret, cardOrganizationName, cardScopes, apiKey, allowedOrigins, organization, cardAPIKey, cardAllowedOrigins }) => {
-  return (
-    <div css={css`
-          display : flex;
-          flex-direction : column;
-          gap: 32px;
-        `}>
-      <h4 className="spectrum-Heading spectrum-Heading--sizeS">{cardClientDetails?.heading}</h4>
-      {cardAPIKey && <CardAPIKey cardClientDetails={cardClientDetails} cardAPIKey={cardAPIKey} apiKey={apiKey} />}
-      {cardClientId && <CardClientId cardClientDetails={cardClientDetails} cardClientId={cardClientId} />}
-      {cardAllowedOrigins && <CardAllowedOrigins cardClientDetails={cardClientDetails} allowedOrigins={allowedOrigins} cardAllowedOrigins={cardAllowedOrigins} />}
-      {cardClientSecret && <CardClientSecret cardClientDetails={cardClientDetails} cardClientSecret={cardClientSecret} />}
-      {cardOrganizationName && <CardOrganizationName cardClientDetails={cardClientDetails} cardOrganizationName={cardOrganizationName} organization={organization} />}
-      {cardScopes && <CardScopes cardClientDetails={cardClientDetails} cardScopes={cardScopes} />}
-    </div>
-  )
-}
-
-const CardAPIKey = ({ cardClientDetails, cardAPIKey, apiKey }) => {
-  return (
-    <ShowCard heading={cardAPIKey?.heading} value={apiKey} />
-  )
-}
-
-const CardAllowedOrigins = ({ cardClientDetails, cardAllowedOrigins, allowedOrigins }) => {
-  return (
-    <ShowCard heading={cardAllowedOrigins?.heading} value={allowedOrigins} />
-  )
-}
-
-const CardClientId = ({ cardClientDetails, cardClientId, apiKey }) => {
-  return (
-    <ShowCard heading={cardClientId?.heading} value={apiKey} />
-  )
-}
-
-const CardClientSecret = ({ cardClientDetails, cardClientSecret }) => {
-  return (
-    <ShowCard heading={cardClientSecret?.heading} isClientSecret buttonLabel={cardClientSecret?.buttonLabel} />
-  )
-}
-
-const CardScopes = ({ cardClientDetails, cardScopes, allowedOrigins }) => {
-  return (
-    <ShowCard heading={cardScopes?.heading} value={allowedOrigins} />
-  )
-}
-
-const CardOrganizationName = ({ cardClientDetails, cardOrganizationName, organization }) => {
-  return (
-    <ShowCard heading={cardOrganizationName?.heading} value={organization} isOraganization />
-  )
-}
-
-const ShowCard = ({ heading, value, isClientSecret, buttonLabel, isOraganization }) => {
-
-  const [isTooltipOpen, setTooltipOpen] = useState(false);
-
-  const handleLeave = () => {
-    setTooltipOpen(false);
-  }
-
-  const handleCopy = (value) => {
-    navigator.clipboard.writeText(value);
-  };
-
-  return (
-    <div
-      css={css`
-          display:flex;
-          flex-direction:column;
-          gap:8px;
-        `}
-    >
-      <p className="spectrum-Body spectrum-Body--sizeS">{heading}</p>
-      <div
-        css={css` 
-            display:flex;
-            align-items: center;
-            gap: 24px; 
-          `}>
-
-
-        {isClientSecret &&
-          <button class="spectrum-Button spectrum-Button--outline spectrum-Button--primary spectrum-Button--sizeM"
-            css={css`
-                cursor : pointer;
-                border: 1px solid var(--spectrum-global-color-gray-400);
-                border-radius: 3px;
-                padding: 3px 6px;
-                height:32px;
-                background :transparent;
-              `}
-          >
-            <span class="spectrum-ActionButton-label">{buttonLabel}</span>
-          </button>}
-
-        {value && <p className="spectrum-Body spectrum-Body--sizeS"
-          css={css`
-              font-family: Source Code Pro,Monaco,monospace;
-              white-space: normal;
-              overflow-wrap: anywhere;
-              max-width: 300px;
-            `}
-        >{value}</p>
-        }
-
-        {!isClientSecret &&
-
-          <div css={css`position:relative;display:${isOraganization ? "none" : "block"}`}>
-            <button className="spectrum-ActionButton spectrum-ActionButton--sizeM"
-              onMouseEnter={() => setTooltipOpen(true)}
-              onMouseLeave={handleLeave}
-              onClick={() => handleCopy(value)}
-              css={css`
-                  border: 1px solid var(--spectrum-global-color-gray-400);
-                  border-radius: 3px;
-                  padding: 3px 6px;
-                `}
-            >
-              {<span className="spectrum-ActionButton-label"><CopyIcon /></span>}
-            </button>
-
-            {isTooltipOpen && (
-              <span
-                className="spectrum-Tooltip spectrum-Tooltip--top is-open"
-                css={css`
-                    position: absolute;
-                    bottom: 25px;
-                    top: unset;
-                    white-space: nowrap;
-                  `}
-              >
-                <span className="spectrum-Tooltip-label">Copy</span>
-                <span className="spectrum-Tooltip-tip"></span>
-              </span>
-            )}
-          </div>
-
-        }
-
-      </div>
-    </div>
-  )
-}
-
-
-export { MyCredential, MyCredentialSide, AccessToken, DevConsoleLink, CardClientDetails, CardClientId, CardClientSecret, CardOrganizationName, CardScopes, CardAPIKey, CardAllowedOrigins };
+export { MyCredential };
