+++ packages/gatsby-theme-aio/src/components/GetCredential/PreviousProject.js
@@ -1,12 +1,18 @@
 import React, { useEffect, useState } from 'react';
 import { css } from "@emotion/react";
 import '@spectrum-css/contextualhelp/dist/index-vars.css';
-import { MAX_MOBILE_WIDTH, MAX_TABLET_SCREEN_WIDTH, MIN_MOBILE_WIDTH, MIN_TABLET_SCREEB_WIDTH, LinkOut, KeyIcon, CopyIcon } from './FormFields';
-import { Picker } from "../Picker";
 import { Toast } from '../Toast';
-import CustomPopover from './CustomPopover';
-import { ActionButton, Tooltip, TooltipTrigger } from '@adobe/react-spectrum';
 import classNames from "classnames";
+import { MAX_TABLET_SCREEN_WIDTH, MIN_MOBILE_WIDTH, LinkOut, KeyIcon, CopyIcon } from './FormFields';
+import { ReturnAccessToken } from './Return/ReturnAccessToken';
+import { ProjectsDropdown } from './Return/ProjectsDropdown';
+import { ReturnDevConsoleLink } from './Return/ReturnDevConsoleLink';
+import { ReturnManageDeveloperConsole } from './Return/ReturnManageDeveloperConsole';
+import ReturnClientDetails from './Return/ReturnClientDetails';
+import { ReturnClientId } from './Return/ReturnClientId';
+import { ReturnClientSecret } from './Return/ReturnClientSecret';
+import { ReturnScopes } from './Return/ReturnScopes';
+import { ReturnProducts } from './Return/ReturnProducts';
 
 const PreviousProject = ({ returnProps, returnFields, productList }) => {
 
@@ -99,7 +105,7 @@ const PreviousProject = ({ returnProps, returnFields, productList }) => {
                   `}
               >
                 <h3 className="spectrum-Heading spectrum-Heading--sizeM">
-                  {/* {formData['CredentialName']} */} {previousProjectsDetails[selectedIndex].name}
+                  {previousProjectsDetails[selectedIndex].name}
                 </h3>
 
                 {productList && <ReturnProducts productList={productList} />}
@@ -251,253 +257,4 @@ const PreviousProject = ({ returnProps, returnFields, productList }) => {
   )
 };
 
-
-const ProjectsDropdown = ({ projectsDropdown, previousProjectsDetails, selectedIndex, setSelectedIndex }) => {
-
-  return (
-
-    <div
-      css={css`
-        display : flex;
-        flex-direction : column;
-        gap:2px;
-    `}
-    >
-      <p className="spectrum-Body spectrum-Body--sizeS" css={css`color: #464646`}>{projectsDropdown?.label} </p>
-      <div
-        css={css`
-                  
-          & > div > .spectrum-Picker {
-            width: 100% !important;
-            height: 20px;
-          }
-
-          & > div > div {
-            width: 22%;
-
-            @media screen and (min-width:${MIN_MOBILE_WIDTH}) and (max-width:${MAX_MOBILE_WIDTH}){
-              width: 82%;
-              left: 15%;
-            }
-
-            @media screen and (min-width:${MIN_TABLET_SCREEB_WIDTH}) and (max-width:${MAX_TABLET_SCREEN_WIDTH}){
-              width: 91%;
-              left: 7%;
-            }
-
-          }
-
-          & > div > .spectrum-Picker-popover > ul > li > div > div {
-            margin : 0 ;
-          }
-
-          & > div > .spectrum-Picker-popover > ul > li > div > div > svg {
-            @media screen and (min-width:${MIN_MOBILE_WIDTH}) and (max-width:${MAX_TABLET_SCREEN_WIDTH}){
-              margin: 3px;
-              padding: 0;
-            }
-          }
-
-            padding: 8px;
-            border-radius: 3px;
-            border: 1px solid #909090 !important;
-            width: 400px;
-
-          ` }
-      >
-        <Picker
-          isQuiet
-          items={previousProjectsDetails.map((organs, k) => {
-            return {
-              title: organs?.name,
-              selected: k === selectedIndex
-            }
-          })}
-          onChange={(index) => {
-            setSelectedIndex(index);
-          }}
-        />
-      </div>
-      <p className="spectrum-Body spectrum-Body--sizeS" css={css`color: #464646`}>{projectsDropdown?.subHeading}</p>
-    </div>
-  )
-}
-
-const ReturnAccessToken = ({ returnAccessToken }) => {
-  return (
-    <div css={css`
-      display : flex;
-      flex-direction : column;
-      gap:16px;
-    `}>
-      <h4 className="spectrum-Heading spectrum-Heading--sizeS">{returnAccessToken?.heading}</h4>
-      <button css={css`width: 180px;`} className="spectrum-Button spectrum-Button--fill spectrum-Button--accent spectrum-Button--sizeM">
-        <span className="spectrum-Button-label">{returnAccessToken?.buttonLabel}</span>
-      </button>
-    </div>
-  )
-}
-
-const ReturnDevConsoleLink = ({ previousProjectsDetails, returnDevConsoleLink, selectedIndex }) => {
-  return (
-    <div css={css`
-      display : flex;
-      flex-direction : column;
-      gap:16px;
-    `}>
-      <h4 className="spectrum-Heading spectrum-Heading--sizeS">{returnDevConsoleLink?.heading}</h4>
-
-      <div css={css`display:flex;`}>
-        <div>
-          <p className="spectrum-Body spectrum-Body--sizeS"
-            css={css`
-            font-family: Source Code Pro,Monaco,monospace;
-            white-space: normal;
-            overflow-wrap: anywhere;
-            max-width: 300px;
-            color: #0265DC;
-          `}
-          >{previousProjectsDetails[selectedIndex].name}</p></div>
-        <div css={
-          css`
-            margin-left:10px;
-            cursor : pointer;
-            @media screen and (min-width:${MIN_MOBILE_WIDTH}) and (max-width:${MAX_TABLET_SCREEN_WIDTH}){
-              display:none;
-            }
-          }`
-        }><LinkOut /></div>
-      </div>
-    </div>
-  )
-}
-
-const ReturnProducts = ({ productList }) => {
-  return (
-
-    <div
-      css={css`
-        display : flex;
-        gap : 10px; 
-        align-items : center;  
-      `}
-    >
-      <ReturnProduct productList={productList} />
-      <CustomPopover productList={productList} />
-    </div>
-  )
-}
-
-const ReturnProduct = ({ productList }) => {
-  return (
-    <>
-      {productList.map((product, index) => {
-        if (index < 2)
-          return (
-            <div
-              css={css`
-                & > button {
-                  border : none !important;
-                }
-              `}
-            >
-              <TooltipTrigger delay={0}>
-                <ActionButton aria-label="Edit Name">
-                  <img
-                    src={product?.icon}
-                    css={css`
-                      width: 35px;
-                      cursor : pointer;
-                    `}
-                  />
-                </ActionButton>
-                <Tooltip>{product?.label}</Tooltip>
-              </TooltipTrigger>
-            </div>
-          )
-      })}
-    </>
-  )
-}
-
-const ReturnManageDeveloperConsole = ({ returnManageDeveloperConsole }) => {
-  return (
-    <div>
-      <div css={css`display:flex;`}>
-        <a href={returnManageDeveloperConsole?.direction} target="_blank" rel="noreferrer"
-          css={css`
-            color:#0265DC;
-          `}
-        >
-          {returnManageDeveloperConsole?.label}
-        </a>
-        <div css={
-          css`
-            margin-left:10px;
-            cursor : pointer;
-
-            @media screen and (min-width:${MIN_MOBILE_WIDTH}) and (max-width:${MAX_TABLET_SCREEN_WIDTH}){
-              display:none;
-            }
-          
-            & > svg > path{
-              fill : #0265DC !important;
-            }
-          `
-        }><LinkOut /></div>
-      </div>
-    </div>
-  )
-}
-
-const ReturnClientDetails = ({ returnClientDetails }) => {
-  return (
-    <div>
-      <ReturnAPIKey returnClientDetails={returnClientDetails} />
-      <ReturnAllowedOrigins returnClientDetails={returnClientDetails} />
-      <ReturnClientId returnClientDetails={returnClientDetails} />
-      <ReturnClientSecret returnClientDetails={returnClientDetails} />
-      <ReturnOrganizationName returnClientDetails={returnClientDetails} />
-      <ReturnScopes returnClientDetails={returnClientDetails} />
-      <h4 className="spectrum-Heading spectrum-Heading--sizeS">{returnClientDetails?.heading}</h4>
-    </div>
-  )
-}
-
-const ReturnAPIKey = ({ returnClientDetails }) => {
-  return (
-    <></>
-  )
-}
-
-const ReturnAllowedOrigins = ({ returnClientDetails }) => {
-  return (
-    <></>
-  )
-}
-
-const ReturnClientId = ({ returnClientDetails }) => {
-  return (
-    <></>
-  )
-}
-
-const ReturnClientSecret = ({ returnClientDetails }) => {
-  return (
-    <></>
-  )
-}
-
-const ReturnScopes = ({ returnClientDetails }) => {
-  return (
-    <></>
-  )
-}
-
-const ReturnOrganizationName = ({ returnClientDetails }) => {
-  return (
-    <></>
-  )
-}
-
-export { PreviousProject, ProjectsDropdown, ReturnAccessToken, ReturnDevConsoleLink, ReturnProducts, ReturnProduct, ReturnManageDeveloperConsole, ReturnClientId, ReturnClientSecret, ReturnScopes, ReturnOrganizationName, ReturnClientDetails, ReturnAPIKey, ReturnAllowedOrigins };
+export { PreviousProject };
