+++ packages/gatsby-theme-aio/src/components/GetCredential/Return/ReturnAPIKey.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const ReturnAPIKey = ({ returnClientDetails }) => {
+    return (
+        <></>
+    )
+}
+
+export { ReturnAPIKey }
