+++ packages/gatsby-theme-aio/src/components/GetCredential/Return/ReturnAllowedOrigins.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const ReturnAllowedOrigins = ({ returnClientDetails }) => {
+  return (
+    <></>
+  )
+}
+
+export { ReturnAllowedOrigins };
