+++ packages/gatsby-theme-aio/src/components/GetCredential/Return/ReturnClientId.js
@@ -1,0 +1,8 @@
+import React from 'react'
+
+const ReturnClientId = ({ returnClientDetails }) => {
+  return (
+    <></>
+  )
+}
+export { ReturnClientId }
