+++ packages/gatsby-theme-aio/src/components/GetCredential/Return/ReturnClientSecret.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const ReturnClientSecret = ({ returnClientDetails }) => {
+  return (
+    <></>
+  )
+}
+
+export { ReturnClientSecret };
