+++ packages/gatsby-theme-aio/src/components/GetCredential/Return/ReturnCredentialDetails.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+const ReturnCredentialDetails = () => { }
+
+export { ReturnCredentialDetails };
