+++ packages/gatsby-theme-aio/src/components/GetCredential/Return/ReturnCustomComp.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+const ReturnCustomComp = ({ retunrSideComp,  returnCustomComp }) => <>{returnCustomComp?.children}</>;
+
+export { ReturnCustomComp };
