+++ packages/gatsby-theme-aio/src/components/GetCredential/Return/ReturnOrganizationName.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const ReturnOrganizationName = ({ returnClientDetails }) => {
+  return (
+    <></>
+  )
+}
+
+export { ReturnOrganizationName };
