+++ packages/gatsby-theme-aio/src/components/GetCredential/Return/ReturnScopes.js
@@ -1,0 +1,10 @@
+import React from 'react'
+
+const ReturnScopes = ({ returnClientDetails }) => {
+  return (
+    <></>
+  )
+}
+
+
+export { ReturnScopes }
