+++ packages/gatsby-theme-aio/src/components/GetCredential/SignIn.js
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import Context from '@adobe/gatsby-theme-aio/src/components/Context';
+import Context from '../Context';
 import { css } from "@emotion/react";
 import classNames from "classnames";
 import { MAX_TABLET_SCREEN_WIDTH, MIN_MOBILE_WIDTH } from './FormFields';
